requirements.txt
nltk
textblob
requests
pyyaml

# Testing dependencies (optional)
# pytest>=7.0.0
